// MT5MG_Conf/src/modules/holiday/types.ts
export type IHoliday = {
    Day: number;
    Description: string;
    Mode: string;
    Month: number;
    WorkFrom: number;
    WorkFromHours: number;
    WorkFromMinutes: number;
    WorkTo: number;
    WorkToHours: number;
    WorkToMinutes: number;
    Year: number;
    Symbols: string[];
  }
  